fix: Fixed notification site filtering not working issues

+++ src/main/java/com/iishanto/jobhunterbackend/domain/usecase/UserJobAccessUseCase.java
@@ -10,5 +10,5 @@ public interface UserJobAccessUseCase {
 
     List<SimpleJobModel> getAppliedJobs(int page, int limit, String query, int siteId);
 
-    List<SimpleJobModel> getAllJobs(int page, int limit, String query, int siteId);
+    List<SimpleJobModel> getAllJobs(int page, int limit, String query, Long siteId);
 }
